Offer UI for Syscoin marketplace
Offer interface including tabs for the reference client. Had to hackify
the offertablemodel (to be fixed) to be able to support my offers vs my
accepted offers... we still need a way to get offers that I accepted
(needs  a solver function and a hook in COfferAccept called
IsAcceptedByMe or something).. seems it was crashing in my offline tests
when I tried to call accept&purchase, via console or via the button
press in the UI tab, but Dan isn't seeing it so I hope its something
localized to my setup. This is the first attempt to merge offer ui code
in.

Created an OfferView to hold all of the offer pages, which may be
customized later... doesnt make sense to have them as one page.