finished single node testing
merge mining confirmed to be working here. Last commit gave the error we
wanted (merkle not correct in auxpow.cpp) and this change pushed MM
blocks through so it confirms it works. Now we need to test services at
the same time as MM blocks, anything that uses the data field.

Set testnet to 1000 and mainnet to something far into the future for
now.