alias model fix (transferred alias hell)
Transferred aliases mean that you may come across an alias some time
earlier in the blockchain and then the UI will display it as yours but
sometime later you have transferred it to someone else, yet the UI
doesn't detect this and displays it as yours still, doesn't even get the
transaction for the transferred one to remove it because its not your
transaction to check... thus the list gets stuck in limbo.

So we have to wait until the blockchain is downloaded before we check
the validity of aliases. Thus I created a refresh button to allow
download of your aliases that is valid only after blockchain is
downloaded.

The logic in the model to download aliases has extra checks to get the
alias from database and ensure that it is your alias (ismine == true).
This will catch retrieving an alias from an older transaction that has
been since transferred (which the database knows about because it has
all transactions now).

I removed any alias notifies... we are always pulling now manually.