Part of last commit
Wasnt using the last tx of alias to determine if it hasnt been
transferred or not... this fixed it. Also any new tx's recieved, if they
have already been processed by the GUI will by skipped. This works
because we have a synced blockchain as a prerequisite to refreshing
aliases on the gui.