aliasdb didnt have nheight stored for live transactions
aliasdb is written to in checkaliasinputs, writing the alias name and
nheight/hash etc.. but nheight seemed to be 0 for new aliases.. which
aliasscan relies on and always returned 0 in the alias search.

I think that what happened was that getvalueofalaistxhash was being
called which reads nHEight from the db which isn't written to yet so new
transactions will always return  here... infact reading the code I saw
that nheight is already set as it is the pindexblock->nheight as used
everywhere else, and on top of that we dont even need vchval to be read
again because vvchArgs should have vchval already from decodealiastx
early on in the function.

Also removed a notify alias update that I was testing in a previous bug
and I DOnt think we need this notifyalias update in the unspent
function.