change getnetworkfee functions to use opcodetype
so that its easier to understand what the input to the functions is,
rather than int... which seemed to not work anyways.

Also added feature to remove transactions on restart of wallet if the
committransaction failed in the wallet... so it won't stay in the ledger
unconfirmed which was annoying.