CheckTransaciton refactoring and coinbase pays too much workaround
seems that if we reset the fees on a coinbase pays too much error the
problem will be fixed, so reset and try once more before throwing
coinbase pays too much error... we know that the fees structures in
memory are not in line with what it was when the transaction was
created, causing fee mismatches... and we know that if we restart wallet
everything works (because fees are reloaded).. so do this in code
instead of forcing user to restart wallet.

Also remove height from service fee calculations... don't need it right
now one less thing to worry about.

Also only add fees to the fee list if > 0... otherwise dont add.