mainnetfork check was wrong, alias fix
HasReachedMainNetForkB2() was only checking for mainnet and that the
hardfork was passed but we weren't exercising the fee checks on cakenet
or testnet which we should have been. If HasReachedMainNetForkB2() is
false it multiplies allowed fees by 800%... it should only do this on
MAINNET and under hardforkB2 block height. On cakenet/testnet should be
good right from the start (block 1)

Alias fix was done to avoid people from entering addresses as alias
names.. although doesn't affect sending to aliases.. it will send to
address first.. just for semantics and clarity it was done.