offerrefund feature
offerrefund rpc call and functionality to work around the concurrency
issue with offeraccept.

If 2 people accept an offer at the same time and there is not enough qty
on that offer to supply both of them then order fills on FIFO basis and
the person that overfills will get refunded automatically.

Its a 2 step process in order to ensure that funds are available to
refund. A 1 step process is possible however this would mean that the
offer owner ensure that they have double the accept amount of coins
available at all times in case of a refund. A 2 step process therefore
is much better to use the coins that were given and not assume account
size is bigger than the amount of accepted coins.