add exclusive whitelist selling, fixed bugs
2 bugs squashed
1) send offeraccept coins in one tx instead of 2 to avoid someone from
sending like 1 sys and claiming an accept
2) update linked offers only if you are the root (not a linked offer)...
linked offers only get updated (qty, title etc) if the root has accepted
the offer or updated itself and thus propagated down to the children.

Added exclusive whitelist feature that allows you to only allow those
owners of the certs in your whitelist of your offer to be able to accept
your offer through a linked accept. This means if this is a linked
accept that gets sent from checkinputs it will check the whitelist to
see if you are able to get "special" access to accept the root offer.
You may apply special pricing in the whitelist entry(the owner) so that
only some people can get those deals and as an owner you may not want
others to resell your offer unless they are specifically added by you in
your whitelist.