Factor in qty to offeraccept checkofferinputs check + fix offerupdate price sync to whitelist offers
Qty wasn't being factored in when checking for how much offeraccept paid
towards the offer. Created a function called GetPrice which takes the
discount into account from the whitelist and is used to get the price of
the offer (if the whitelist entry is passed in).. thus you should check
that a whitelist entry exists for you and pass it in to getprice to get
accurate price info.

The offerupdate wasn't updating all of the children offer prices
properly so I scale all children prices by the ratio of how much the
owner offer price changed by. Scaling is efficient because with
whitelisting you'd have to somehow get the whitelist entry of every
single children offer recursively to recalculate how much the price
should be if the parent changed.