Allow price discovery of 1 USD using SYS_FEE alias
using a special alias we can peg he value of offers aswell as fees to
use USD so that when sys value changes over time prices will be pegged
to USD to avoid volatility issues. Fees also are pegged to fractions of
USD using this alias. Allows for scalable services in the long run.